+++ cg/meta/store/mip.py
@@ -1,11 +1,44 @@
 """ MIP specific functionality for storing files in Houskeeper """
+import logging
 from pathlib import Path
 import ruamel.yaml
 
+from cg.constants import HK_TAGS
 from cg.exc import (
     AnalysisNotFinishedError,
+    BundleAlreadyAddedError,
 )
-from cg.meta.store.base import build_bundle
+from cg.meta.store.base import (
+    get_files,
+    add_new_analysis,
+    reset_case_action,
+)
+
+
+LOG = logging.getLogger(__name__)
+
+
+def gather_files_and_bundle_in_housekeeper(config_stream, hk_api, status):
+    """Function to gather files and bundle in housekeeper"""
+    bundle_data = add_analysis(config_stream)
+
+    results = hk_api.add_bundle(bundle_data)
+    if results is None:
+        raise BundleAlreadyAddedError("bundle already added")
+
+    bundle_obj, version_obj = results
+
+    case_obj = status.family(bundle_obj.name)
+
+    reset_case_action(case_obj)
+    new_analysis = add_new_analysis(bundle_data, case_obj, status, version_obj)
+    version_date = version_obj.created_at.date()
+
+    LOG.info("new bundle added: %s, version %s", bundle_obj.name, version_date)
+    hk_api.include(version_obj)
+    hk_api.add_commit(bundle_obj, version_obj)
+
+    return new_analysis
 
 
 def add_analysis(config_stream):
@@ -72,3 +105,18 @@ def _get_sample_analysis_type(data: dict) -> list:
         {"id": sample_id, "type": analysis_type}
         for sample_id, analysis_type in data["analysis_type"].items()
     ]
+
+
+def build_bundle(config_data: dict, sampleinfo_data: dict, deliverables: dict) -> dict:
+    """Create a new bundle for."""
+
+    pipeline = config_data["samples"][0]["type"]
+    pipeline_tag = HK_TAGS[pipeline]
+
+    data = {
+        "name": config_data["case"],
+        "created": sampleinfo_data["date"],
+        "pipeline_version": sampleinfo_data["version"],
+        "files": get_files(deliverables, pipeline_tag),
+    }
+    return data
